refactor(courtboard): extract ToastHost and LoadingPlaceholder components
- Create src/courtboard/components/ directory
- Extract ToastHost to components/ToastHost.jsx
- Extract LoadingPlaceholder to components/LoadingPlaceholder.jsx
- Add imports in main.jsx, remove inline definitions
